requirements.txt
python-telegram-bot>=20.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
schedule>=1.0.0
modelcontextprotocol>=0.1.0
google-search-results>=2.4.0